+++ uefi-test-runner/src/boot/memory.rs
@@ -1,15 +1,48 @@
-use uefi::table::boot::{AllocateType, BootServices, MemoryType};
-
 use alloc::vec::Vec;
+use uefi::boot;
+use uefi::mem::memory_map::{MemoryMap, MemoryMapMut, MemoryType};
+use uefi::table::boot::{AllocateType, BootServices};
 
 pub fn test(bt: &BootServices) {
     info!("Testing memory functions");
+
+    test_allocate_pages_freestanding();
+    test_allocate_pool_freestanding();
 
     allocate_pages(bt);
     vec_alloc();
     alloc_alignment();
 
     memory_map(bt);
+}
+
+fn test_allocate_pages_freestanding() {
+    let num_pages = 1;
+    let ptr =
+        boot::allocate_pages(AllocateType::AnyPages, MemoryType::LOADER_DATA, num_pages).unwrap();
+    let addr = ptr.as_ptr() as usize;
+    assert_eq!(addr % 4096, 0, "Page pointer is not page-aligned");
+
+    // Verify the page can be written to.
+    {
+        let ptr = ptr.as_ptr();
+        unsafe { ptr.write_volatile(0xff) };
+        unsafe { ptr.add(4095).write_volatile(0xff) };
+    }
+
+    unsafe { boot::free_pages(ptr, num_pages) }.unwrap();
+}
+
+fn test_allocate_pool_freestanding() {
+    let ptr = boot::allocate_pool(MemoryType::LOADER_DATA, 10).unwrap();
+
+    // Verify the allocation can be written to.
+    {
+        let ptr = ptr.as_ptr();
+        unsafe { ptr.write_volatile(0xff) };
+        unsafe { ptr.add(9).write_volatile(0xff) };
+    }
+    unsafe { boot::free_pool(ptr) }.unwrap();
 }
 
 fn allocate_pages(bt: &BootServices) {
@@ -63,41 +96,37 @@ fn alloc_alignment() {
 fn memory_map(bt: &BootServices) {
     info!("Testing memory map functions");
 
-    // Ensure that the memory map is freed after each iteration (on drop).
-    // Otherwise, we will have an OOM.
-    for _ in 0..200000 {
-        let mut memory_map = bt
-            .memory_map(MemoryType::LOADER_DATA)
-            .expect("Failed to retrieve UEFI memory map");
+    let mut memory_map = bt
+        .memory_map(MemoryType::LOADER_DATA)
+        .expect("Failed to retrieve UEFI memory map");
 
-        memory_map.sort();
+    memory_map.sort();
 
-        // Collect the descriptors into a vector
-        let descriptors = memory_map.entries().copied().collect::<Vec<_>>();
+    // Collect the descriptors into a vector
+    let descriptors = memory_map.entries().copied().collect::<Vec<_>>();
 
-        // Ensured we have at least one entry.
-        // Real memory maps usually have dozens of entries.
-        assert!(!descriptors.is_empty(), "Memory map is empty");
+    // Ensured we have at least one entry.
+    // Real memory maps usually have dozens of entries.
+    assert!(!descriptors.is_empty(), "Memory map is empty");
 
-        let mut curr_value = descriptors[0];
+    let mut curr_value = descriptors[0];
 
-        for value in descriptors.iter().skip(1) {
-            if value.phys_start <= curr_value.phys_start {
-                panic!("memory map sorting failed");
-            }
-            curr_value = *value;
+    for value in descriptors.iter().skip(1) {
+        if value.phys_start <= curr_value.phys_start {
+            panic!("memory map sorting failed");
         }
-
-        // This is pretty much a basic sanity test to ensure returned memory
-        // isn't filled with random values.
-        let first_desc = descriptors[0];
-
-        #[cfg(target_arch = "x86_64")]
-        {
-            let phys_start = first_desc.phys_start;
-            assert_eq!(phys_start, 0, "Memory does not start at address 0");
-        }
-        let page_count = first_desc.page_count;
-        assert!(page_count != 0, "Memory map entry has size zero");
+        curr_value = *value;
     }
+
+    // This is pretty much a basic sanity test to ensure returned memory
+    // isn't filled with random values.
+    let first_desc = descriptors[0];
+
+    #[cfg(target_arch = "x86_64")]
+    {
+        let phys_start = first_desc.phys_start;
+        assert_eq!(phys_start, 0, "Memory does not start at address 0");
+    }
+    let page_count = first_desc.page_count;
+    assert!(page_count != 0, "Memory map entry has size zero");
 }